libunvmed/cmds: return EEIXST if given @cid exists
Application can provide a specific @cid value within [0, qsize) for the
corresponding SQ.  In this case, libunvmed tries to allocate the
specific value of @cid from the bitmap, and if it fails, it should
return an error to notify the user application to let it retry or
failfast.  This patch returned EEIXST if the given @cid has already been
allocated, and updated the library documentation also.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>