unvmed-cmds: fix to grab usq lock before alloc cmd
Currently, there is a timing issue that allocated and not submitted @cmd
instance is not canceled in reset thread. Because of this, the reset
thread can not be done.

        Thread 1 (passthru)             Thread 2 (reset)
        ------------------              ----------------
        cmd allocation
                                        lock(usq)
                                        cancel cmds (only submitted cmds)
                                        wait to @usq->nr_cmds == 0 (hang)
                                        unlock(usq)
        lock(usq) (hang)
        cmd post

This patch grabs @usq->lock before allocating @cmd instance, so that all
commands are submitted right after allocated.

At now, we apply this only to `unvme_passthru` for testing. After
verification, we have to spread out to other cli commands.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>