libunvmed: mark @cmd to be completed in unvmed_cmd_cmpl
unvmed_cmd_cmpl() tries to _complete_ the given @cmd instance for two
different cases: (1) CLI commands (WAKEUP_ON_CQE) and (2) FIO commands
(VCQ-based).  In case of CLI commands, we can simply _complete_ the
command instance by putting @cqe to @cmd->cqe and mark them completed.
But, in case of FIO commands, @cmd instance should be pushed to the
@vcq in the following sequence so that we should mark the @cmd instance
to be completed soon.

Without this, the following issue may happen:

	FIO thread			Reset thread
	----------			------------
	cqe = unvmed_vcq_pop();
					if (cmd->state == SUBMITTED)
						unvmed_put_cqe(fake cqe)
							unvmed_vcq_push();
	__unvmed_cmd_cmpl(cmd)

The issue is mainly about the @cmd->state.  Even if @cqe has been
fetched from @vcq normally, reset thread has recognized the @cmd state
as SUBMITTED, which means @cqe has not arrived yet, and it should be
cancelled.  This misleads fake @cqe to be pushed to @vcq again and it
will crash.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>