libunvmed: set @usq|ucq->q to NULL after discard
In commit 715169fd544f, we fixed to discard @usq|ucq->q. When discard
those, libvfn just set 0-value to the instance, not to free it. Because
of this, the daemon process cannot know about the @usq|ucq->q is alive
or not.

For clearity, this patch set the @usq|ucq->q to NULL explicitly, so that
the daemon knows the instance is alive or not.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>