unvmed-cmds: fix to check @sq is enabled
If the given @usq is disabled, it means the device is in reset progress.
CLI tries to update S/Q doorbell at this time, segmentation fault will
be raised because of the @usq->q is set to NULL. To avoid this, check
the given @usq is enabled before updating S/Q doorbell.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>