libunvmed: fix get/put scheme for @usq and @ucq
If we don't check @refcnt > 0 in the while-loop, we can't simply catch
the situation that an instance was just about to be freed by refcnt ==
0.  In this case, we can't catch the following situation:

	thread A			thread B
	--------			--------
	unvmed_sq_put()
	  usq->refcnt-- = 0
					unvmed_sq_get()
					  0 = load(usq->refcnt)
					  cas(usq->refcnt, 0, 1)
	  free(usq);

So, we should _firstly_ check the atomically loaded @refcnt value is
zero or not.  If zero, it means that this instance was just about to be
freed by other thread context and we should not return the instance.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>