unvme-cmds: fix to check the given @usq is enabled
If fio is running, fio grabs @usq instances and increases @usq->refcnt.
Because of this, even if reset is triggered, the @usq instances will not
be freed before fio release them. In this situation, if user tries to
create I/O Submission queue, it fails with errno EEXIT, because the
unvmed considers that the @usq instance is alive and already enabled.

To avoid this, this patch fixes to check the @usq is enabled, rather
than @usq is not NULL.

Signed-off-by: Jonghwi Jeong <jongh2.jeong@samsung.com>