unvme-cli: add --shmem-size=<bytes> to `unvme add`
Added a new option --shmem-size=<bytes> to pre-allocate a huge size of
memory buffer in the unvmed process with being shared among different
processes for exteranl usages.  For example, other application process
might want to manage the data buffer management rather than calling
`malloc`, `mfree`, `mread`, and `mwrite` everytime they want to manage
the memory chunks which might cause inefficiency.

If user gives --shmem-size with value, unvmed will create a shmem object
into the system named /unvmed-<underscored_bdf> which can be found at
/dev/shm/.  Application can open the shmem file descriptor and manage
the pre-allocated memory buffer freely based on the pre-mapped IOVA, so
application does not need to map the vaddr to iova again.

Maybe later, we can convert it to the hugepage, but for now, this is
kind of sufficient way to expose a shared memory to the other processes.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>