README: remove --disable-tls when building fio
It is a mechanism that allows each thread in a multi-threaded program to
have its own independent copy of a variable.  This ensures that data
stored in TLS is private to the thread that created or modified it — no
other thread can access or interfere with it.  TLS is commonly used to
avoid race conditions and eliminate the need for locks when managing
per-thread state, such as error codes, logging IDs, or temporary
buffers.

If TLS is disabled, fio will create a TLS key by `pthread_create_key`
in `fio_clock_init()`, but never deleted since fio considers the TLS key
for clock only based on one-time-living process.  If we don't enable
tls, the key will never be deleted and maximum number of TLS will
exceed with a return value of `pthread_create_key()`.

	EAGAIN PTHREAD_KEYS_MAX keys are already allocated

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>